feat: add basic createTask flow.
Created Task object: object containing all properties required
Created TaskManager: this class will manage the main functionality of the aplication. Contains the list of task and his id. I added the feature to add tasks.
Created TaskStatus: enum with the three states of a task (todo, in progress and done)